refactor: render `Main` component inside `Calendar` factory class
* create new constant of view state
