+++ apps/calendar/src/constants/view.ts
@@ -1,0 +1,9 @@
+import { ViewType } from '@src/model';
+
+export const VIEW_TYPE: {
+  [key: string]: ViewType;
+} = {
+  MONTH: 'month',
+  WEEK: 'week',
+  DAY: 'day',
+};
